Edit function setting the url of example (#12)
