+++ .idea/misc.xml
@@ -12,6 +12,7 @@
         <entry key="..\:/AndroidStudioProjects/AndroidStudy/app/src/main/res/layout/item_list.xml" value="0.3333333333333333" />
         <entry key="..\:/AndroidStudioProjects/AndroidStudy/app/src/main/res/layout/loadinganiexam_main.xml" value="0.16666666666666666" />
         <entry key="..\:/AndroidStudioProjects/AndroidStudy/app/src/main/res/layout/servicexam_main.xml" value="0.1" />
+        <entry key="..\:/AndroidStudioProjects/AndroidStudy/app/src/main/res/layout/setting_dialog.xml" value="0.28125" />
         <entry key="..\:/AndroidStudioProjects/AndroidStudy/app/src/main/res/layout/spinnerexam_main.xml" value="0.4466666666666667" />
         <entry key="..\:/AndroidStudioProjects/AndroidStudy/app/src/main/res/layout/webviewexam_main.xml" value="0.3333333333333333" />
         <entry key="..\:/AndroidStudioProjects/AndroidStudy/app/src/main/res/xml/aaa.xml" value="0.1" />
