+++ app/src/main/res/layout/item_list.xml
@@ -10,18 +10,20 @@
         android:id="@+id/item_default"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:orientation="horizontal">
+        android:weightSum="5">
 
         <ImageView
             android:id="@+id/iv_profile"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
+            android:orientation="horizontal"
             android:src="@mipmap/ic_launcher"/>
 
         <LinearLayout
             android:gravity="center_vertical"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
+            android:layout_weight="1.2"
             android:orientation="vertical">
 
             <TextView
@@ -45,10 +47,24 @@
                 android:layout_height="wrap_content"
                 android:textSize="20dp"
                 android:text="1"/>
+        </LinearLayout>
 
+        <LinearLayout
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            android:gravity="center_vertical"
+            android:layout_weight="3.8">
+
+            <ImageButton
+                android:id="@+id/btn_setting"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:src="@drawable/outline_settings_24"/>
         </LinearLayout>
 
     </LinearLayout>
+
+
 
     <!-- View for expend -->
     <LinearLayout
